finalisation : TODO Test

+++ app/src/main/java/com/example/pboguet/macaveonline/Utils/Adapters/LieuAchatAdapter.java
@@ -40,7 +40,7 @@ public class LieuAchatAdapter extends ArrayAdapter<LieuAchat> {
             nom = (TextView) v.findViewById(R.id.nom);
             if(id != null)
             {
-                id.setText(Long.toString(lieu.getId()));
+                id.setText(Integer.toString(lieu.getId()));
                 nom.setText(lieu.getNom());
             }
         }
